third_party/blink/web_tests/platform/generic/fast/dom/MutationObserver/database-callback-delivery-expected.txt

Testing mutations are delivered following Database transaction callbacks.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].attributeName is "foo"


PASS successfullyParsed is true

TEST COMPLETE
